+++ tests/test_workflows.py
@@ -7,7 +7,7 @@ import geopandas as gpd
 from pytest import MonkeyPatch
 from pytest_mock import MockerFixture
 
-from dist_s1.data_models.output_models import ProductDirectoryData
+from dist_s1.data_models.output_models import DistS1ProductDirectory
 from dist_s1.data_models.runconfig_model import RunConfigData
 from dist_s1.rio_tools import check_profiles_match, open_one_profile
 from dist_s1.workflows import (
@@ -95,7 +95,7 @@ def test_dist_s1_sas_workflow(
     run_dist_s1_sas_workflow(config)
 
     product_data = config.product_data_model
-    product_data_golden = ProductDirectoryData.from_product_path(test_opera_golden_dummy_dataset)
+    product_data_golden = DistS1ProductDirectory.from_product_path(test_opera_golden_dummy_dataset)
 
     assert product_data == product_data_golden